Ensuring backwards compatibility when switching between true and false states of nightlyRun feature.
