+++ config/env.yaml
@@ -27,4 +27,4 @@ Parameters:
   ImportCurrentSSOConfiguration: false
   UpgradeFromVersionLessThanV303: false # Should be set to true, if upgrading from solution version 3.0.2/earlier
   EnableNightlyRun: true
-  NightlyRunRemediationMode: "AUTOREMEDIATE" # If EnableNightlyrun is set to true, this should be set to AUTOREMEDIATE or NOTIFY
+  NightlyRunRemediationMode: "NOTIFY" # If EnableNightlyrun is set to true, this should be set to AUTOREMEDIATE or NOTIFY
