+++ lib/lambda-functions/package.json
@@ -7,17 +7,17 @@
   },
   "keywords": [],
   "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.99.0",
-    "@aws-sdk/client-identitystore": "^3.99.0",
-    "@aws-sdk/client-s3": "^3.99.0",
-    "@aws-sdk/client-sfn": "^3.99.0",
-    "@aws-sdk/client-sns": "^3.99.0",
-    "@aws-sdk/client-sqs": "^3.99.0",
-    "@aws-sdk/client-ssm": "^3.99.0",
-    "@aws-sdk/client-sso-admin": "^3.99.0",
-    "@aws-sdk/credential-providers": "^3.99.0",
-    "@aws-sdk/lib-dynamodb": "^3.99.0",
-    "@aws-sdk/util-dynamodb": "^3.99.0",
+    "@aws-sdk/client-dynamodb": "^3.100.0",
+    "@aws-sdk/client-identitystore": "^3.100.0",
+    "@aws-sdk/client-s3": "^3.100.0",
+    "@aws-sdk/client-sfn": "^3.100.0",
+    "@aws-sdk/client-sns": "^3.100.0",
+    "@aws-sdk/client-sqs": "^3.100.0",
+    "@aws-sdk/client-ssm": "^3.100.0",
+    "@aws-sdk/client-sso-admin": "^3.100.0",
+    "@aws-sdk/credential-providers": "^3.100.0",
+    "@aws-sdk/lib-dynamodb": "^3.100.0",
+    "@aws-sdk/util-dynamodb": "^3.100.0",
     "@aws-sdk/util-waiter": "^3.78.0",
     "ajv": "^8.11.0",
     "json-diff": "^0.7.4",
