+++ lib/stacks/pipelineStageStacks/sso-import-artefacts-part3.ts
@@ -1,5 +1,5 @@
 /*
-This is used for any post clean up operations required
+This is used for any post clean up / completions operations required for SSO import current config/ nightlyRun use cases
 */
 import { Stack, StackProps } from "aws-cdk-lib";
 import { PolicyStatement } from "aws-cdk-lib/aws-iam";
@@ -9,14 +9,11 @@ import {
   PhysicalResourceId,
 } from "aws-cdk-lib/custom-resources";
 import { Construct } from "constructs";
-import { v4 as uuidv4 } from "uuid";
 import { BuildConfig } from "../../build/buildConfig";
 
 function name(buildConfig: BuildConfig, resourcename: string): string {
   return buildConfig.Environment + "-" + resourcename;
 }
-
-const generateRandomString = uuidv4().toString().split("-")[0];
 
 export class SSOImportArtefactsPart3 extends Stack {
   constructor(
@@ -27,20 +24,20 @@ export class SSOImportArtefactsPart3 extends Stack {
   ) {
     super(scope, id, props);
 
-    /** Nightly run post clean up operations */
-    if (buildConfig.Parameters.EnableNightlyRun === true) {
-      /**
-       * Enable nightlyRun rule to trigger the logic, as the import lambda's are
-       * already created We use non deterministic physicalResourceId to
-       * guarantee that the nightly run enablement is idempotent
-       */
-      const enablenightlyRunPolicyStatement = new PolicyStatement({
-        resources: [
-          `arn:aws:events:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:rule/${buildConfig.Environment}-nightlyRunRule`,
-        ],
-        actions: ["events:EnableRule"],
-      });
+    /**
+     * For nightlyRun use case, if the feature is enabled we enable the rule,
+     * else we disable the rule
+     */
+    /** Construct the required permissions policy for this */
+    const nightlyRunMgmtPolicy = new PolicyStatement({
+      resources: [
+        `arn:aws:events:${buildConfig.PipelineSettings.SSOServiceAccountRegion}:${buildConfig.PipelineSettings.SSOServiceAccountId}:rule/${buildConfig.Environment}-nightlyRunRule`,
+      ],
+      actions: ["events:EnableRule", "events:DisableRule"],
+    });
 
+    if (buildConfig.Parameters.EnableNightlyRun) {
+      /** Nightly run is enabled, so we enable the rule */
       new AwsCustomResource(this, name(buildConfig, "enableNightlyRunRule"), {
         onCreate: {
           service: "EventBridge",
@@ -48,12 +45,28 @@ export class SSOImportArtefactsPart3 extends Stack {
           parameters: {
             Name: name(buildConfig, "nightlyRunRule"),
           },
-          physicalResourceId: PhysicalResourceId.of(generateRandomString),
+          physicalResourceId: PhysicalResourceId.of(
+            `${buildConfig.Environment}-nightlyRunRule-enablement`
+          ),
           region: buildConfig.PipelineSettings.SSOServiceAccountRegion,
         },
-        policy: AwsCustomResourcePolicy.fromStatements([
-          enablenightlyRunPolicyStatement,
-        ]),
+        policy: AwsCustomResourcePolicy.fromStatements([nightlyRunMgmtPolicy]),
+      });
+    } else {
+      /** Nightly run is disabled/unknowmn state, so we disable the rule */
+      new AwsCustomResource(this, name(buildConfig, "disableNightlyRunRule"), {
+        onCreate: {
+          service: "EventBridge",
+          action: "disableRule",
+          parameters: {
+            Name: name(buildConfig, "nightlyRunRule"),
+          },
+          physicalResourceId: PhysicalResourceId.of(
+            `${buildConfig.Environment}-nightlyRunRule-disablement`
+          ),
+          region: buildConfig.PipelineSettings.SSOServiceAccountRegion,
+        },
+        policy: AwsCustomResourcePolicy.fromStatements([nightlyRunMgmtPolicy]),
       });
     }
   }
