+++ package.json
@@ -28,12 +28,12 @@
     "@types/aws-lambda": "^8.10.97",
     "@types/jest": "^27.5.1",
     "@types/js-yaml": "^4.0.5",
-    "@types/node": "^17.0.35",
+    "@types/node": "^17.0.36",
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.26.0",
     "@typescript-eslint/parser": "^5.26.0",
     "aws-cdk": "^2.25.0",
-    "esbuild": "^0.14.39",
+    "esbuild": "^0.14.42",
     "eslint": "^8.16.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-security": "^1.5.0",
@@ -47,7 +47,7 @@
   },
   "dependencies": {
     "aws-cdk-lib": "^2.25.0",
-    "constructs": "^10.1.18",
+    "constructs": "^10.1.21",
     "js-yaml": "^4.1.0",
     "source-map-support": "^0.5.21"
   }
